+++ frontend/app/registro/camionero.tsx
@@ -15,8 +15,8 @@ import { Ionicons } from '@expo/vector-icons';
 import SuccessModal from "../_components/SuccessModal";
 
 const BACKEND_URL = process.env.EXPO_PUBLIC_BACKEND_URL;
-const licencias = ["AM", "A1", "A2", "A", "B", "C1", "C", "C1+E", "C+E", "D1", "D+E","D1+E","D"];
-const licencias_backend = ["AM", "A1", "A2", "A", "B", "C1", "C", "C1_E", "C_E", "D1", "D_E","D1_E","D"];
+const licencias = ["AM", "A1", "A2", "A", "B", "C1", "C", "C1+E", "C+E", "D1", "D+E", "D1+E", "D"];
+const licencias_backend = ["AM", "A1", "A2", "A", "B", "C1", "C", "C1_E", "C_E", "D1", "D_E", "D1_E", "D"];
 
 const CamioneroScreen = () => {
   const router = useRouter();
@@ -52,37 +52,37 @@ const CamioneroScreen = () => {
   };
 
   const handlePickImage = async () => {
-      const base64Image = await pickImageAsync();
-      if (base64Image) {
-        setFormData((prevState) => ({ 
-          ...prevState, 
-          foto: base64Image.base64,
-          fotoUri: base64Image.uri
-        }));
-      }
-    };
-    
-    const pickImageAsync = async (): Promise<{ uri: string; base64: string } | null> => {
-      try {
-        const result = await ImagePicker.launchImageLibraryAsync({
-          mediaTypes: ImagePicker.MediaTypeOptions.Images,
-          allowsEditing: true,
-          aspect: [1, 1],
-          quality: 1,
-          base64: true,
-        });
-    
-        if (!result.canceled && result.assets && result.assets.length > 0) {
-          const image = result.assets[0]; 
-          return { uri: image.uri, base64: image.base64 }; 
-        } else {
-          return null;
-        }
-      } catch (error) {
-        console.error("Error picking image: ", error);
+    const base64Image = await pickImageAsync();
+    if (base64Image) {
+      setFormData((prevState) => ({
+        ...prevState,
+        foto: base64Image.base64,
+        fotoUri: base64Image.uri
+      }));
+    }
+  };
+
+  const pickImageAsync = async (): Promise<{ uri: string; base64: string } | null> => {
+    try {
+      const result = await ImagePicker.launchImageLibraryAsync({
+        mediaTypes: ImagePicker.MediaTypeOptions.Images,
+        allowsEditing: true,
+        aspect: [1, 1],
+        quality: 1,
+        base64: true,
+      });
+
+      if (!result.canceled && result.assets && result.assets.length > 0) {
+        const image = result.assets[0];
+        return { uri: image.uri, base64: image.base64 };
+      } else {
         return null;
       }
-    };
+    } catch (error) {
+      console.error("Error picking image: ", error);
+      return null;
+    }
+  };
 
   const handleRegister = async () => {
     const licenciasBackend = formData.licencias.map((licencia) => licencias_backend[licencias.indexOf(licencia)]);
@@ -131,11 +131,11 @@ const CamioneroScreen = () => {
       login(responseLogin.data, token);
 
       setSuccessModalVisible(true);
-        setTimeout(() => {
-          setSuccessModalVisible(false);
-          router.replace("/");
-        }, 1000);
-      
+      setTimeout(() => {
+        setSuccessModalVisible(false);
+        router.replace("/");
+      }, 1000);
+
     } catch (error) {
       console.error('Error en la solicitud', error);
       if (axios.isAxiosError(error) && error.response && error.response.data && error.response.data.message) {
@@ -148,17 +148,17 @@ const CamioneroScreen = () => {
         setErrorMessage('Error desconocido');
       }
     }
-      
+
   };
 
   // Render input function
   const renderInput = (label, field, icon, keyboardType = "default", secureTextEntry = false, multiline = false) => (
     <View style={{ width: '90%', marginBottom: 15 }}>
       <Text style={{ fontSize: 16, color: colors.secondary, marginLeft: 8, marginBottom: -6, backgroundColor: colors.white, alignSelf: 'flex-start', paddingHorizontal: 5, zIndex: 1 }}>{label}</Text>
-      <View style={{ flexDirection: "row", alignItems: "center", borderWidth: 1, borderColor: colors.mediumGray, borderRadius: 8, paddingHorizontal: 10, backgroundColor: colors.white }}>      
+      <View style={{ flexDirection: "row", alignItems: "center", borderWidth: 1, borderColor: colors.mediumGray, borderRadius: 8, paddingHorizontal: 10, backgroundColor: colors.white }}>
         {icon}
         <TextInput
-          style={{ flex: 1, height: multiline ? 80 : 40, paddingLeft: 8, outline:"none", textAlignVertical: multiline ? 'top' : 'center' }}
+          style={{ flex: 1, height: multiline ? 80 : 40, paddingLeft: 8, outline: "none", textAlignVertical: multiline ? 'top' : 'center' }}
           keyboardType={keyboardType}
           secureTextEntry={secureTextEntry}
           multiline={multiline}
@@ -194,16 +194,16 @@ const CamioneroScreen = () => {
         <View style={styles.cardContainer}>
 
           <TouchableOpacity style={styles.backButton} onPress={() => router.push('/registro')}>
-              <Ionicons name="arrow-back" size={30} color="#0b4f6c" />
+            <Ionicons name="arrow-back" size={30} color="#0b4f6c" />
           </TouchableOpacity>
 
           <Text style={styles.title}>Registro como Camionero</Text>
 
           {/* Foto de perfil */}
           <View style={{ alignItems: "center", marginBottom: 20, marginTop: 10 }}>
-            <Image 
-              source={formData.fotoUri ? { uri: formData.fotoUri } : defaultProfileImage}  
-              style={{ width: 100, height: 100, borderRadius: 50, marginBottom: 8, borderWidth: 1, borderColor: colors.mediumGray }} 
+            <Image
+              source={formData.fotoUri ? { uri: formData.fotoUri } : defaultProfileImage}
+              style={{ width: 100, height: 100, borderRadius: 50, marginBottom: 8, borderWidth: 1, borderColor: colors.mediumGray }}
             />
 
             {!formData.foto ? (
@@ -217,7 +217,7 @@ const CamioneroScreen = () => {
                   <MaterialIcons name="cached" size={20} color={colors.white} style={{ marginRight: 8 }} />
                   <Text style={globalStyles.buttonText}>Cambiar</Text>
                 </TouchableOpacity>
-                
+
                 <TouchableOpacity onPress={() => setFormData({ ...formData, fotoUri: null })} style={[globalStyles.button, { backgroundColor: colors.red, flexDirection: "row", alignItems: "center", justifyContent: "center", width: 120, paddingHorizontal: 15 }]}>
                   <FontAwesome5 name="trash" size={18} color={colors.white} style={{ marginRight: 8 }} />
                   <Text style={globalStyles.buttonText}>Borrar</Text>
@@ -238,27 +238,27 @@ const CamioneroScreen = () => {
           {/* Campos del formulario específicos de camionero */}
           {renderInput("DNI", "dni", <FontAwesome5 name="address-card" size={20} color={colors.primary} />)}
           <View style={styles.inputContainer}>
-              <Text style={{ color: colors.secondary, fontSize: 16, marginRight: 10, flexDirection: "row", alignItems: "center", marginBottom: 10 }}>
-                Selecciona tus licencias:
-              </Text>
-              <MultiSelector 
-                value={formData.licencias}
-                onChange={(value) => handleInputChange("licencias", value)}
-                options={licencias}
-                colors={colors} 
-              />
-            </View>
+            <Text style={{ color: colors.secondary, fontSize: 16, marginRight: 10, flexDirection: "row", alignItems: "center", marginBottom: 10 }}>
+              Selecciona tus licencias:
+            </Text>
+            <MultiSelector
+              value={formData.licencias}
+              onChange={(value) => handleInputChange("licencias", value)}
+              options={licencias}
+              colors={colors}
+            />
+          </View>
 
           <View style={styles.inputContainer}>
             <Text style={{ color: colors.secondary, fontSize: 16, marginRight: 10, flexDirection: "row", alignItems: "center", marginBottom: 10 }}>
               Disponibilidad:
             </Text>
-            <Selector 
-              value={formData.disponibilidad} 
-              onChange={(value) => handleInputChange("disponibilidad", value)} 
-              options={["NACIONAL", "INTERNACIONAL"]} 
-              colors={colors} 
-              globalStyles={globalStyles} 
+            <Selector
+              value={formData.disponibilidad}
+              onChange={(value) => handleInputChange("disponibilidad", value)}
+              options={["NACIONAL", "INTERNACIONAL"]}
+              colors={colors}
+              globalStyles={globalStyles}
             />
           </View>
 
@@ -269,11 +269,11 @@ const CamioneroScreen = () => {
             <Text style={{ color: colors.secondary, fontSize: 16, marginRight: 10, flexDirection: "row", alignItems: "center", marginBottom: 10 }}>
               ¿Tiene CAP (Certificado de Aptitud Profesional)?:
             </Text>
-            <BooleanSelector 
-              value={formData.tieneCAP} 
-              onChange={(value) => handleInputChange("tieneCAP", value)} 
-              colors={colors} 
-              globalStyles={globalStyles} 
+            <BooleanSelector
+              value={formData.tieneCAP}
+              onChange={(value) => handleInputChange("tieneCAP", value)}
+              colors={colors}
+              globalStyles={globalStyles}
             />
           </View>
 
@@ -290,11 +290,11 @@ const CamioneroScreen = () => {
             <Text style={{ color: colors.secondary, fontSize: 16, marginRight: 10, flexDirection: "row", alignItems: "center", marginBottom: 10 }}>
               ¿Eres autónomo/a?:
             </Text>
-            <BooleanSelector 
-              value={formData.isAutonomo} 
-              onChange={(value) => handleInputChange("isAutonomo", value)} 
-              colors={colors} 
-              globalStyles={globalStyles} 
+            <BooleanSelector
+              value={formData.isAutonomo}
+              onChange={(value) => handleInputChange("isAutonomo", value)}
+              colors={colors}
+              globalStyles={globalStyles}
             />
           </View>
 
@@ -304,11 +304,11 @@ const CamioneroScreen = () => {
               <Text style={{ color: colors.secondary, fontSize: 16, marginRight: 10, flexDirection: "row", alignItems: "center", marginBottom: 10 }}>
                 Selecciona tus tarjetas:
               </Text>
-              <MultiSelector 
+              <MultiSelector
                 value={formData.tarjetas}
                 onChange={(value) => handleInputChange("tarjetas", value)}
                 options={["VTC", "VC", "MSL", "MDP"]}
-                colors={colors} 
+                colors={colors}
               />
             </View>
           )}
