+++ frontend/app/workinprogress.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
-import { useAuth } from '../../contexts/AuthContext';
+import { useAuth } from '../contexts/AuthContext';
 
 const WorkInProgressScreen: React.FC = () => {
     const { logout } = useAuth();
